Fix imm to be i128

+++ jolt-core/src/zkvm/instruction/auipc.rs
@@ -29,27 +29,27 @@ impl InstructionFlags for AUIPC {
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<AUIPC> {
     fn to_lookup_operands(&self) -> (u64, u128) {
         let (pc, imm) = LookupQuery::<WORD_SIZE>::to_instruction_inputs(self);
-        (0, (pc as i128 + imm as i128) as u128)
+        (0, (pc as i128 + imm) as u128)
     }
 
     fn to_lookup_index(&self) -> u128 {
         LookupQuery::<WORD_SIZE>::to_lookup_operands(self).1
     }
 
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         match WORD_SIZE {
             #[cfg(test)]
             8 => (
                 self.instruction.address as u8 as u64,
-                self.instruction.operands.imm as u8 as u64 as i64,
+                self.instruction.operands.imm as u8 as u64 as i128,
             ),
             32 => (
                 self.instruction.address as u32 as u64,
-                self.instruction.operands.imm as u32 as u64 as i64,
+                self.instruction.operands.imm as u32 as u64 as i128,
             ),
             64 => (
                 self.instruction.address,
-                self.instruction.operands.imm as i64,
+                self.instruction.operands.imm as i128,
             ),
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
@@ -61,7 +61,7 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<AUIPC> {
             #[cfg(test)]
             8 => (pc as i8).overflowing_add(imm as i8).0 as u8 as u64,
             32 => (pc as i32).overflowing_add(imm as i32).0 as u32 as u64,
-            64 => (pc as i64).overflowing_add(imm).0 as u64,
+            64 => (pc as i64).overflowing_add(imm as i64).0 as u64,
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
     }
