+++ jolt-core/src/zkvm/instruction/mod.rs
@@ -16,7 +16,7 @@ pub trait InstructionLookup<const WORD_SIZE: usize> {
 pub trait LookupQuery<const WORD_SIZE: usize> {
     /// Returns a tuple of the instruction's inputs. If the instruction has only one input,
     /// one of the tuple values will be 0.
-    fn to_instruction_inputs(&self) -> (u64, i64);
+    fn to_instruction_inputs(&self) -> (u64, i128);
 
     /// Returns a tuple of the instruction's lookup operands. By default, these are the
     /// same as the instruction inputs returned by `to_instruction_inputs`, but in some cases
@@ -160,7 +160,7 @@ macro_rules! define_rv32im_trait_impls {
         }
 
         impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RV32IMCycle {
-            fn to_instruction_inputs(&self) -> (u64, i64) {
+            fn to_instruction_inputs(&self) -> (u64, i128) {
                 match self {
                     RV32IMCycle::NoOp => (0, 0),
                     $(
