+++ jolt-core/src/zkvm/instruction/slt.rs
@@ -26,18 +26,18 @@ impl InstructionFlags for SLT {
 }
 
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<SLT> {
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         match WORD_SIZE {
             #[cfg(test)]
             8 => (
                 self.register_state.rs1 as u8 as u64,
-                self.register_state.rs2 as u8 as i64,
+                self.register_state.rs2 as u8 as i128,
             ),
             32 => (
                 self.register_state.rs1 as u32 as u64,
-                self.register_state.rs2 as u32 as i64,
+                self.register_state.rs2 as u32 as i128,
             ),
-            64 => (self.register_state.rs1, self.register_state.rs2 as i64),
+            64 => (self.register_state.rs1, self.register_state.rs2 as i128),
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
     }
@@ -48,7 +48,7 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<SLT> {
             #[cfg(test)]
             8 => ((x as i8) < (y as i8)).into(),
             32 => ((x as i32) < (y as i32)).into(),
-            64 => ((x as i64) < y).into(),
+            64 => ((x as i64) < (y as i64)).into(),
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
     }
