+++ jolt-core/src/zkvm/instruction/virtual_advice.rs
@@ -25,7 +25,7 @@ impl InstructionFlags for VirtualAdvice {
 }
 
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualAdvice> {
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         (0, 0)
     }
