+++ jolt-core/src/zkvm/instruction/virtual_pow2.rs
@@ -27,7 +27,7 @@ impl InstructionFlags for VirtualPow2 {
 }
 
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualPow2> {
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         match WORD_SIZE {
             #[cfg(test)]
             8 => (self.register_state.rs1 as u8 as u64, 0),
