+++ jolt-core/src/zkvm/instruction/virtual_pow2w.rs
@@ -27,7 +27,7 @@ impl InstructionFlags for VirtualPow2W {
 }
 
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualPow2W> {
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         // Only use rs1 value
         (self.register_state.rs1, 0)
     }
