+++ tracer/src/instruction/bltu.rs
@@ -19,7 +19,7 @@ impl BLTU {
     fn exec(&self, cpu: &mut Cpu, _: &mut <BLTU as RISCVInstruction>::RAMAccess) {
         if cpu.unsigned_data(cpu.x[self.operands.rs1]) < cpu.unsigned_data(cpu.x[self.operands.rs2])
         {
-            cpu.pc = (self.address as i64 + self.operands.imm) as u64;
+            cpu.pc = (self.address as i64 + self.operands.imm as i64) as u64;
         }
     }
 }
